file

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,13 +7,13 @@
 
 int main(void)
 {
-	int num, count;
+	long int num, count;
 
 	for (count = 2 ; num > 1 ; count++)
 	{
 		while (num % count == 0)
 		{
-			printf("%d", count);
+			printf("%ld", count);
 			num = num / count;
 		}
 	}
